refactor(a2a): split target configuration component
