+++ src/app/src/pages/redteam/setup/components/Targets/CustomTargetConfiguration.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 
 import { Alert, AlertContent, AlertDescription } from '@app/components/ui/alert';
 import { Button } from '@app/components/ui/button';
@@ -8,16 +8,8 @@ import {
   CollapsibleContent,
   CollapsibleTrigger,
 } from '@app/components/ui/collapsible';
-import { HelperText } from '@app/components/ui/helper-text';
 import { Input } from '@app/components/ui/input';
 import { Label } from '@app/components/ui/label';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@app/components/ui/select';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@app/components/ui/tooltip';
 import Prism from '@app/lib/prism';
 import { cn } from '@app/lib/utils';
@@ -27,8 +19,6 @@ import {
   ChevronDown,
   Cloud,
   Code2,
-  Eye,
-  EyeOff,
   FileCode2,
   Server,
   Terminal,
@@ -60,51 +50,6 @@ interface ProviderConfig {
   configDescription: string;
 }
 
-type A2AAuthConfig = {
-  type?: 'bearer' | 'basic' | 'api_key' | 'oauth' | 'no_auth';
-  token?: string;
-  username?: string;
-  password?: string;
-  keyName?: string;
-  value?: string;
-  placement?: 'header' | 'query';
-  grantType?: 'client_credentials' | 'password';
-  tokenUrl?: string;
-  clientId?: string;
-  clientSecret?: string;
-  scopes?: string[] | string;
-};
-
-type A2AProviderConfig = Record<string, unknown> & {
-  url?: string;
-  agentCardUrl?: string;
-  mode?: 'auto' | 'send' | 'stream';
-  tenant?: string;
-  protocolVersion?: string;
-  auth?: A2AAuthConfig;
-};
-
-const A2A_STRUCTURED_CONFIG_KEYS = new Set(['url', 'agentCardUrl', 'auth']);
-
-const asA2AConfig = (config?: ProviderOptions['config']): A2AProviderConfig =>
-  (config ?? {}) as A2AProviderConfig;
-
-const getA2AAdvancedConfig = (config?: ProviderOptions['config']): Record<string, unknown> =>
-  Object.fromEntries(
-    Object.entries(config ?? {}).filter(([key]) => !A2A_STRUCTURED_CONFIG_KEYS.has(key)),
-  );
-
-const getA2AStructuredConfig = (config?: ProviderOptions['config']): Record<string, unknown> =>
-  Object.fromEntries(
-    Object.entries(config ?? {}).filter(([key]) => A2A_STRUCTURED_CONFIG_KEYS.has(key)),
-  );
-
-const parseA2AScopes = (value: string): string[] =>
-  value
-    .split(',')
-    .map((scope) => scope.trim())
-    .filter(Boolean);
-
 const highlightJSON = (code: string): string => {
   try {
     const grammar = Prism?.languages?.json;
@@ -117,105 +62,8 @@ const highlightJSON = (code: string): string => {
   }
 };
 
-const SecretInput = ({
-  id,
-  label,
-  value,
-  onChange,
-  placeholder,
-  helperText,
-  required,
-}: {
-  id: string;
-  label: string;
-  value: string;
-  onChange: (value: string) => void;
-  placeholder?: string;
-  helperText?: string;
-  required?: boolean;
-}) => {
-  const [isVisible, setIsVisible] = useState(false);
-
-  return (
-    <div className="space-y-2">
-      <Label htmlFor={id}>
-        {label}
-        {required && <span className="text-destructive"> *</span>}
-      </Label>
-      <div className="relative">
-        <Input
-          id={id}
-          type={isVisible ? 'text' : 'password'}
-          value={value}
-          onChange={(e) => onChange(e.target.value)}
-          placeholder={placeholder}
-          className="pr-10"
-          autoComplete="off"
-        />
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <Button
-              type="button"
-              variant="ghost"
-              size="icon"
-              className="absolute right-0 top-0 h-full px-3 hover:bg-transparent"
-              onClick={() => setIsVisible((current) => !current)}
-              aria-label={isVisible ? `Hide ${label.toLowerCase()}` : `Show ${label.toLowerCase()}`}
-            >
-              {isVisible ? <EyeOff className="size-4" /> : <Eye className="size-4" />}
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent>
-            {isVisible ? `Hide ${label.toLowerCase()}` : `Show ${label.toLowerCase()}`}
-          </TooltipContent>
-        </Tooltip>
-      </div>
-      {helperText && <p className="text-sm text-muted-foreground">{helperText}</p>}
-    </div>
-  );
-};
-
 const getProviderConfig = (providerType?: string): ProviderConfig => {
   switch (providerType) {
-    case 'a2a':
-      return {
-        title: 'A2A Provider',
-        icon: <Server className="size-5 text-primary" />,
-        targetIdLabel: 'Provider ID',
-        targetIdPlaceholder: 'a2a or a2a:https://agent.example.com/a2a/v1',
-        helpText: (
-          <>
-            Agent2Agent (A2A) HTTP+JSON endpoint configuration. Promptfoo sends test prompts as A2A
-            messages and reads final messages, task artifacts, or streaming events.
-          </>
-        ),
-        docUrl: 'https://www.promptfoo.dev/docs/providers/a2a/',
-        examples: {
-          title: 'A2A Provider Examples',
-          items: [
-            { code: 'a2a:https://agent.example.com/a2a/v1', description: 'Endpoint shorthand' },
-            { code: 'a2a', description: 'Use config.url or config.agentCardUrl' },
-          ],
-        },
-        configExample: {
-          url: 'https://agent.example.com/a2a/v1',
-          agentCardUrl: 'https://agent.example.com/.well-known/agent-card.json',
-          mode: 'auto',
-          tenant: 'optional-tenant',
-          protocolVersion: '1.0',
-          auth: {
-            type: 'bearer',
-            token: '{{A2A_API_KEY}}',
-          },
-          polling: {
-            enabled: true,
-            intervalMs: 1000,
-            timeoutMs: 300000,
-          },
-        },
-        configDescription: 'A2A endpoint, Agent Card discovery, headers, and task polling',
-      };
-
     case 'python':
       return {
         title: 'Python Provider',
@@ -559,32 +407,12 @@ const CustomTargetConfiguration = ({
 }: CustomTargetConfigurationProps) => {
   const [targetId, setTargetId] = useState(selectedTarget.id?.replace('file://', '') || '');
   const [docsExpanded, setDocsExpanded] = useState(false);
-  const [advancedConfigError, setAdvancedConfigError] = useState<string | null>(null);
-  const initializedA2ATargetId = useRef<string | null>(null);
 
   const config = useMemo(() => getProviderConfig(providerType), [providerType]);
-  const a2aConfig = asA2AConfig(selectedTarget.config);
 
   useEffect(() => {
     setTargetId(selectedTarget.id?.replace('file://', '') || '');
   }, [selectedTarget.id]);
-
-  useEffect(() => {
-    if (providerType === 'a2a') {
-      const targetKey = selectedTarget.id ?? '';
-
-      if (initializedA2ATargetId.current === targetKey) {
-        return;
-      }
-
-      initializedA2ATargetId.current = targetKey;
-      setRawConfigJson(JSON.stringify(getA2AAdvancedConfig(selectedTarget.config), null, 2));
-      setAdvancedConfigError(null);
-      return;
-    }
-
-    initializedA2ATargetId.current = null;
-  }, [providerType, selectedTarget.config, selectedTarget.id, setRawConfigJson]);
 
   const handleTargetIdChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const value = e.target.value;
@@ -603,37 +431,7 @@ const CustomTargetConfiguration = ({
     updateCustomTarget('id', idToSave);
   };
 
-  const handleA2AAdvancedConfigChange = (content: string) => {
-    setRawConfigJson(content);
-
-    try {
-      const parsedConfig = JSON.parse(content);
-
-      if (
-        typeof parsedConfig !== 'object' ||
-        parsedConfig === null ||
-        Array.isArray(parsedConfig)
-      ) {
-        setAdvancedConfigError('Advanced configuration must be a JSON object');
-        return;
-      }
-
-      setAdvancedConfigError(null);
-      updateCustomTarget('config', {
-        ...parsedConfig,
-        ...getA2AStructuredConfig(selectedTarget.config),
-      });
-    } catch {
-      setAdvancedConfigError('Invalid JSON configuration');
-    }
-  };
-
   const handleConfigChange = (content: string) => {
-    if (providerType === 'a2a') {
-      handleA2AAdvancedConfigChange(content);
-      return;
-    }
-
     setRawConfigJson(content);
     try {
       const parsedConfig = JSON.parse(content);
@@ -650,271 +448,11 @@ const CustomTargetConfiguration = ({
         const parsed = JSON.parse(rawConfigJson);
         const formatted = JSON.stringify(parsed, null, 2);
         setRawConfigJson(formatted);
-        if (providerType === 'a2a') {
-          updateCustomTarget('config', {
-            ...parsed,
-            ...getA2AStructuredConfig(selectedTarget.config),
-          });
-          setAdvancedConfigError(null);
-        } else {
-          updateCustomTarget('config', parsed);
-        }
+        updateCustomTarget('config', parsed);
       } catch {
         // Error state is already shown via bodyError prop
       }
     }
-  };
-
-  const updateA2AField = (field: string, value: unknown) => {
-    updateCustomTarget(field, value);
-  };
-
-  const updateA2AAuth = (auth: A2AAuthConfig | undefined) => {
-    updateCustomTarget('auth', auth);
-  };
-
-  const updateA2AAuthField = (field: keyof A2AAuthConfig, value: unknown) => {
-    updateA2AAuth({
-      ...(a2aConfig.auth ?? {}),
-      [field]: value,
-    });
-  };
-
-  const handleA2AAuthTypeChange = (value: string) => {
-    switch (value) {
-      case 'bearer':
-        updateA2AAuth({ type: 'bearer', token: '' });
-        break;
-      case 'basic':
-        updateA2AAuth({ type: 'basic', username: '', password: '' });
-        break;
-      case 'api_key':
-        updateA2AAuth({ type: 'api_key', keyName: 'X-API-Key', placement: 'header', value: '' });
-        break;
-      case 'oauth':
-        updateA2AAuth({
-          type: 'oauth',
-          grantType: 'client_credentials',
-          tokenUrl: '',
-          clientId: '',
-          clientSecret: '',
-          scopes: [],
-        });
-        break;
-      default:
-        updateA2AAuth(undefined);
-    }
-  };
-
-  const handleA2AOAuthGrantTypeChange = (grantType: 'client_credentials' | 'password') => {
-    const currentAuth = a2aConfig.auth ?? {};
-
-    if (grantType === 'password') {
-      updateA2AAuth({
-        ...currentAuth,
-        type: 'oauth',
-        grantType,
-        username: currentAuth.username ?? '',
-        password: currentAuth.password ?? '',
-      });
-      return;
-    }
-
-    const { username: _username, password: _password, ...authWithoutPassword } = currentAuth;
-    updateA2AAuth({
-      ...authWithoutPassword,
-      type: 'oauth',
-      grantType,
-    });
-  };
-
-  const renderA2AAuthFields = () => {
-    const auth = a2aConfig.auth;
-
-    return (
-      <div className="mt-6 space-y-4 rounded-lg border border-border p-4">
-        <div>
-          <h4 className="font-medium">Authorization</h4>
-          <p className="text-sm text-muted-foreground">
-            Configure common A2A authentication without manually editing headers.
-          </p>
-        </div>
-
-        <div className="space-y-2">
-          <Label htmlFor="a2a-auth-type">Authentication Type</Label>
-          <Select value={auth?.type ?? 'no_auth'} onValueChange={handleA2AAuthTypeChange}>
-            <SelectTrigger id="a2a-auth-type">
-              <SelectValue placeholder="Select authentication type" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="no_auth">No Auth</SelectItem>
-              <SelectItem value="bearer">Bearer Token</SelectItem>
-              <SelectItem value="basic">Basic Auth</SelectItem>
-              <SelectItem value="api_key">API Key</SelectItem>
-              <SelectItem value="oauth">OAuth 2.0</SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-
-        {auth?.type === 'bearer' && (
-          <SecretInput
-            id="a2a-bearer-token"
-            label="Bearer Token"
-            value={auth.token ?? ''}
-            onChange={(value) => updateA2AAuthField('token', value)}
-            placeholder="{{A2A_API_KEY}}"
-            helperText="Sent as an Authorization: Bearer header."
-            required
-          />
-        )}
-
-        {auth?.type === 'basic' && (
-          <div className="grid gap-4 md:grid-cols-2">
-            <div className="space-y-2">
-              <Label htmlFor="a2a-basic-username">
-                Username <span className="text-destructive">*</span>
-              </Label>
-              <Input
-                id="a2a-basic-username"
-                value={auth.username ?? ''}
-                onChange={(e) => updateA2AAuthField('username', e.target.value)}
-              />
-            </div>
-            <SecretInput
-              id="a2a-basic-password"
-              label="Password"
-              value={auth.password ?? ''}
-              onChange={(value) => updateA2AAuthField('password', value)}
-              required
-            />
-          </div>
-        )}
-
-        {auth?.type === 'api_key' && (
-          <div className="grid gap-4 md:grid-cols-3">
-            <div className="space-y-2">
-              <Label htmlFor="a2a-api-key-placement">Placement</Label>
-              <Select
-                value={auth.placement ?? 'header'}
-                onValueChange={(value) => updateA2AAuthField('placement', value)}
-              >
-                <SelectTrigger id="a2a-api-key-placement">
-                  <SelectValue placeholder="Select placement" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="header">Header</SelectItem>
-                  <SelectItem value="query">Query Parameter</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="a2a-api-key-name">
-                Key Name <span className="text-destructive">*</span>
-              </Label>
-              <Input
-                id="a2a-api-key-name"
-                value={auth.keyName ?? 'X-API-Key'}
-                onChange={(e) => updateA2AAuthField('keyName', e.target.value)}
-                placeholder={auth.placement === 'query' ? 'api_key' : 'X-API-Key'}
-              />
-            </div>
-            <SecretInput
-              id="a2a-api-key-value"
-              label="API Key Value"
-              value={auth.value ?? ''}
-              onChange={(value) => updateA2AAuthField('value', value)}
-              placeholder="{{A2A_API_KEY}}"
-              required
-            />
-          </div>
-        )}
-
-        {auth?.type === 'oauth' && (
-          <div className="space-y-4">
-            <div className="grid gap-4 md:grid-cols-2">
-              <div className="space-y-2">
-                <Label htmlFor="a2a-oauth-grant-type">Grant Type</Label>
-                <Select
-                  value={auth.grantType ?? 'client_credentials'}
-                  onValueChange={(value) =>
-                    handleA2AOAuthGrantTypeChange(value as 'client_credentials' | 'password')
-                  }
-                >
-                  <SelectTrigger id="a2a-oauth-grant-type">
-                    <SelectValue placeholder="Select grant type" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="client_credentials">Client Credentials</SelectItem>
-                    <SelectItem value="password">Username & Password</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="a2a-oauth-token-url">Token URL</Label>
-                <Input
-                  id="a2a-oauth-token-url"
-                  value={auth.tokenUrl ?? ''}
-                  onChange={(e) => updateA2AAuthField('tokenUrl', e.target.value)}
-                  placeholder="https://agent.example.com/oauth/token"
-                />
-                <p className="text-sm text-muted-foreground">
-                  Optional when the Agent Card exposes an auth realm that supports discovery.
-                </p>
-              </div>
-            </div>
-            <div className="grid gap-4 md:grid-cols-2">
-              <div className="space-y-2">
-                <Label htmlFor="a2a-oauth-client-id">Client ID</Label>
-                <Input
-                  id="a2a-oauth-client-id"
-                  value={auth.clientId ?? ''}
-                  onChange={(e) => updateA2AAuthField('clientId', e.target.value)}
-                  placeholder="{{A2A_CLIENT_ID}}"
-                />
-              </div>
-              <SecretInput
-                id="a2a-oauth-client-secret"
-                label="Client Secret"
-                value={auth.clientSecret ?? ''}
-                onChange={(value) => updateA2AAuthField('clientSecret', value)}
-                placeholder="{{A2A_CLIENT_SECRET}}"
-              />
-            </div>
-            {auth.grantType === 'password' && (
-              <div className="grid gap-4 md:grid-cols-2">
-                <div className="space-y-2">
-                  <Label htmlFor="a2a-oauth-username">
-                    Username <span className="text-destructive">*</span>
-                  </Label>
-                  <Input
-                    id="a2a-oauth-username"
-                    value={auth.username ?? ''}
-                    onChange={(e) => updateA2AAuthField('username', e.target.value)}
-                  />
-                </div>
-                <SecretInput
-                  id="a2a-oauth-password"
-                  label="Password"
-                  value={auth.password ?? ''}
-                  onChange={(value) => updateA2AAuthField('password', value)}
-                  required
-                />
-              </div>
-            )}
-            <div className="space-y-2">
-              <Label htmlFor="a2a-oauth-scopes">Scopes</Label>
-              <Input
-                id="a2a-oauth-scopes"
-                value={Array.isArray(auth.scopes) ? auth.scopes.join(', ') : (auth.scopes ?? '')}
-                onChange={(e) => updateA2AAuthField('scopes', parseA2AScopes(e.target.value))}
-                placeholder="agent:invoke, tasks:read"
-              />
-              <p className="text-sm text-muted-foreground">Comma-separated OAuth scopes.</p>
-            </div>
-          </div>
-        )}
-      </div>
-    );
   };
 
   return (
@@ -955,75 +493,24 @@ const CustomTargetConfiguration = ({
           </p>
         </div>
 
-        {providerType === 'a2a' && (
-          <>
-            <div className="mt-6 space-y-4 rounded-lg border border-border p-4">
-              <div>
-                <h4 className="font-medium">Connection</h4>
-                <p className="text-sm text-muted-foreground">
-                  Start with an Agent Card when available. Use an endpoint URL only to override
-                  discovery or connect directly.
-                </p>
-              </div>
-              <div className="grid gap-4 md:grid-cols-2">
-                <div className="space-y-2">
-                  <Label htmlFor="a2a-agent-card-url">Agent Card URL</Label>
-                  <Input
-                    id="a2a-agent-card-url"
-                    value={a2aConfig.agentCardUrl ?? ''}
-                    onChange={(e) => updateA2AField('agentCardUrl', e.target.value)}
-                    placeholder="https://agent.example.com/.well-known/agent-card.json"
-                  />
-                  <p className="text-sm text-muted-foreground">
-                    Recommended. Used to discover the HTTP+JSON endpoint, tenant, and protocol
-                    version.
-                  </p>
-                </div>
-                <div className="space-y-2">
-                  <Label htmlFor="a2a-url">A2A Endpoint URL</Label>
-                  <Input
-                    id="a2a-url"
-                    value={a2aConfig.url ?? ''}
-                    onChange={(e) => updateA2AField('url', e.target.value)}
-                    placeholder="https://agent.example.com/a2a/v1"
-                  />
-                  <p className="text-sm text-muted-foreground">
-                    Optional override. Used for /message:send, /message:stream, and task polling.
-                  </p>
-                </div>
-              </div>
-            </div>
-
-            {renderA2AAuthFields()}
-          </>
-        )}
-
         {/* Custom Configuration Section */}
         <div className="mt-6 space-y-2">
           <div className="flex items-center justify-between">
-            <Label htmlFor="config-json">
-              {providerType === 'a2a' ? 'Advanced Configuration (JSON)' : 'Configuration (JSON)'}
-            </Label>
+            <Label htmlFor="config-json">Configuration (JSON)</Label>
             <Tooltip>
               <TooltipTrigger asChild>
                 <Button
                   variant="ghost"
                   size="sm"
                   onClick={handleFormatJson}
-                  disabled={
-                    !rawConfigJson.trim() ||
-                    !!bodyError ||
-                    (providerType === 'a2a' && !!advancedConfigError)
-                  }
+                  disabled={!rawConfigJson.trim() || !!bodyError}
                   className="h-7 px-2"
                 >
                   <AlignLeft className="size-4" />
                   <span className="ml-1 text-xs">Format</span>
                 </Button>
               </TooltipTrigger>
-              <TooltipContent>
-                {bodyError || advancedConfigError ? 'Fix JSON errors first' : 'Format JSON'}
-              </TooltipContent>
+              <TooltipContent>{bodyError ? 'Fix JSON errors first' : 'Format JSON'}</TooltipContent>
             </Tooltip>
           </div>
 
@@ -1076,7 +563,7 @@ const CustomTargetConfiguration = ({
           <div
             className={cn(
               'overflow-hidden rounded-lg border',
-              bodyError || advancedConfigError ? 'border-destructive' : 'border-border',
+              bodyError ? 'border-destructive' : 'border-border',
             )}
           >
             <div className="flex items-center justify-between border-b border-border bg-muted/50 px-3 py-1.5">
@@ -1088,33 +575,7 @@ const CustomTargetConfiguration = ({
                 onValueChange={handleConfigChange}
                 highlight={highlightJSON}
                 padding={12}
-                placeholder={
-                  providerType === 'a2a'
-                    ? JSON.stringify(
-                        {
-                          mode: 'auto',
-                          tenant: 'optional-tenant',
-                          protocolVersion: '1.0',
-                          polling: {
-                            enabled: true,
-                            intervalMs: 1000,
-                            timeoutMs: 300000,
-                          },
-                          headers: {
-                            'X-Custom-Header': '{{value}}',
-                          },
-                          message: {
-                            role: 'ROLE_USER',
-                            parts: [{ text: '{{prompt}}' }],
-                          },
-                          transformResponse:
-                            'return result.output || result.message || JSON.stringify(result.raw);',
-                        },
-                        null,
-                        2,
-                      )
-                    : JSON.stringify(config.configExample, null, 2)
-                }
+                placeholder={JSON.stringify(config.configExample, null, 2)}
                 style={{
                   fontFamily: 'ui-monospace, "Fira Code", monospace',
                   fontSize: 13,
@@ -1124,18 +585,13 @@ const CustomTargetConfiguration = ({
             </div>
           </div>
 
-          {bodyError || advancedConfigError ? (
+          {bodyError ? (
             <Alert variant="destructive" className="py-2">
               <AlertCircle className="size-4" />
               <AlertContent>
-                <AlertDescription>{bodyError || advancedConfigError}</AlertDescription>
+                <AlertDescription>{bodyError}</AlertDescription>
               </AlertContent>
             </Alert>
-          ) : providerType === 'a2a' ? (
-            <HelperText>
-              Optional JSON merged with the fields above. Use it for headers, custom message
-              templates, mode, tenant, protocol version, polling, timeouts, and transformResponse.
-            </HelperText>
           ) : (
             <p className="text-sm text-muted-foreground">{config.configDescription}</p>
           )}
